split nodes into seperate files
